+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Links basics
+* Require auth for access links controller (except visit action)
+* Scaffold links
+* Add rspec-rails
+
 # Initial
 * Rails application
 * Add rubocop